pcmrecord -W 0.8: resync changes, report block_drops
Add -Y command line arg to specify the resync tolerance (default is 2
seconds) for wsprdaemon mode. If the first packet of a wav file
arrives outside the resync tolerance window, the next wav file will
start at the :59 -> :00 transition to resync.

If the block_drops counter changes, a warning message will be emitted.

If the RTP timestamp or sequence number isn't as expected, a warning
message will be emitted, but this will not force a wav file resync.