WIP: pcmrecord add PPS sync start
Starts the file ok, but they're not the correct length. Follow on
files in -W mode appear to start correctly.

Out of order detection is broken. Only works for IQ and only for
floats. File length isn't correct. Need to split packet at end of
file. Also need to include some metadata.

And it needs to be tested.