Hacks for testing at Northern Utah
Reduce stderr output in BPSK sync mode if not verbose.
Measure phase shift over multiple sample to try and detect the PPS
edge when using heavy filtering. Not really appropriate for use with a
standard BPSK PPS injector. This is all pretty hacky, and I'm going to
back most of it out in the next commit.